[REFACTO] Ajoute un package conversation
- Afin de refléter le métier et séparer les différents cas utilisateurs
- … et faire de même dans les tests
