+++ src/conversation/question_utilisateur.py
@@ -1,25 +1,24 @@
 import uuid
-from uuid import UUID
 
-from typing import NamedTuple, Union
+from typing import Union, NamedTuple
 
 from adaptateurs import AdaptateurBaseDeDonnees
 from adaptateurs.chiffrement import AdaptateurChiffrement
+
 from adaptateurs.journal import (
     TypeEvenement,
     DonneesInteractionCreee,
     DonneesViolationDetectee,
     AdaptateurJournal,
 )
-from schemas.albert import ReponseQuestion
-from schemas.retour_utilisatrice import (
-    Interaction,
-    RetourUtilisatrice,
-    DonneesCreationRetourUtilisateur,
+from conversation.conversation import (
+    QuestionUtilisateur,
+    ResultatConversation,
+    ResultatInteractionEnErreur,
     Conversation,
-    TagPositif,
-    TagNegatif,
+    Interaction,
 )
+from schemas.albert import ReponseQuestion
 from schemas.type_utilisateur import TypeUtilisateur
 from services.service_albert import ServiceAlbert
 
@@ -29,31 +28,6 @@ class ConfigurationQuestion(NamedTuple):
     adaptateur_base_de_donnees: AdaptateurBaseDeDonnees
     adaptateur_journal: AdaptateurJournal
     adaptateur_chiffrement: AdaptateurChiffrement
-
-
-class ResultatConversation:
-    def __init__(
-        self,
-        id_conversation: uuid.UUID,
-        reponse_question: ReponseQuestion,
-        interaction: Interaction,
-        id_interaction: str,
-    ):
-        self.id_interaction = id_interaction
-        self.reponse_question = reponse_question
-        self.interaction = interaction
-        self.id_conversation = id_conversation
-
-
-class ResultatInteractionEnErreur:
-    def __init__(self, e: Exception):
-        self.message_mqc = "Erreur lors de l’appel à Albert"
-        self.erreur = str(e)
-
-
-class QuestionUtilisateur(NamedTuple):
-    question: str
-    conversation: uuid.UUID | None = None
 
 
 def pose_question_utilisateur(
@@ -127,56 +101,3 @@ def __cree_interaction(
         reponse_question=reponse_question, retour_utilisatrice=None, id=uuid.uuid4()
     )
     return interaction, reponse_question
-
-
-def ajoute_retour_utilisatrice(
-    donnees_ajout_retour: DonneesCreationRetourUtilisateur,
-    adaptateur_base_de_donnees: AdaptateurBaseDeDonnees,
-) -> RetourUtilisatrice | None:
-    if donnees_ajout_retour.id_conversation is not None:
-        id_conversation = donnees_ajout_retour.id_conversation
-
-        conversation = adaptateur_base_de_donnees.recupere_conversation(
-            id_conversation=uuid.UUID(id_conversation)
-        )
-
-        interaction = list(
-            filter(
-                lambda i: str(i.id) == donnees_ajout_retour.id_interaction,
-                conversation.interactions,
-            )
-        )[0]
-        retour = donnees_ajout_retour.retour
-        if len(conversation.interactions) < 2:
-            retour.tags = [
-                t
-                for t in retour.tags
-                if t != TagNegatif.Conversation and t != TagPositif.Conversation
-            ]  # type: ignore[assignment]
-    else:
-        interaction = adaptateur_base_de_donnees.recupere_interaction(
-            uuid.UUID(donnees_ajout_retour.id_interaction)
-        )
-        retour = donnees_ajout_retour.retour
-
-    if not interaction:
-        return None
-
-    interaction.retour_utilisatrice = retour
-    adaptateur_base_de_donnees.sauvegarde_interaction(interaction)
-    return interaction.retour_utilisatrice
-
-
-def supprime_retour_utilisatrice(
-    identifiant_interaction: UUID, adaptateur_base_de_donnees: AdaptateurBaseDeDonnees
-) -> None:
-    interaction = adaptateur_base_de_donnees.recupere_interaction(
-        identifiant_interaction
-    )
-
-    if not interaction:
-        return None
-
-    interaction.retour_utilisatrice = None
-    adaptateur_base_de_donnees.sauvegarde_interaction(interaction)
-    return None
