+++ src/schemas/retour_utilisatrice.py
@@ -1,13 +1,10 @@
-import uuid
 from datetime import datetime
 from enum import StrEnum, auto
 from typing import Generic, Literal, Optional, TypeVar, Union
-from uuid import UUID
 
 from pydantic import BaseModel, Field
 
 from adaptateurs.horloge import AdaptateurHorloge
-from schemas.albert import ReponseQuestion
 
 
 class TagPositif(StrEnum):
@@ -59,34 +56,6 @@ class RetourPositif(AbstractRetourUtilisatrice[TagPositif]):
 
 
 type RetourUtilisatrice = Union[RetourNegatif, RetourPositif]
-
-
-class Interaction(BaseModel):
-    reponse_question: ReponseQuestion
-    retour_utilisatrice: Optional[RetourUtilisatrice] = None
-    date_creation: datetime = Field(default_factory=AdaptateurHorloge.maintenant)
-    id: UUID
-
-
-class Conversation:
-    def __init__(self, interaction: Interaction):
-        self.id_conversation = uuid.uuid4()
-        self._interactions = [interaction]
-
-    @property
-    def interactions(self) -> list[Interaction]:
-        self._interactions.sort(key=lambda i: i.date_creation, reverse=True)
-        return self._interactions
-
-    def ajoute_interaction(self, interaction):
-        self._interactions.append(interaction)
-
-    @staticmethod
-    def hydrate(id_conversation: uuid.UUID, les_interactions: list[Interaction]):
-        conversation = Conversation(les_interactions[0])
-        conversation.id_conversation = id_conversation
-        conversation._interactions.extend(les_interactions[1:])
-        return conversation
 
 
 class DonneesCreationRetourUtilisateur(BaseModel):
